# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inngest App API",
    "version": "1.0.0",
    "description": "API documentation for the Inngest app"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format: Bearer {token}"
      }
    },
    "schemas": {
...
      "WorkflowUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the workflow instance is enabled for use by the user"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the workflow is currently active/scheduled to run"
          },
          "canBeScheduled": {
            "type": "boolean"
          },
          "cronExpressions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timezone": {
            "type": "string"
          },
          "input": {
            "type": "object"
          },
          "credentialIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/workflows/{id}": {
      "get": {
        "summary": "Get a workflow",
        "description": "Returns a single workflow for the authenticated user.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single workflow."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Workflow not found."
          },
          "500": {
            "description": "Error getting workflow."
          }
        }
      },
      "put": {
        "summary": "Update a workflow",
        "description": "Updates a single workflow for the authenticated user.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated workflow."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Error updating workflow."
          }
        }
      },
      "delete": {
        "summary": "Delete a workflow",
        "description": "Deletes a single workflow for the authenticated user.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Error deleting workflow."
          }
        }
      }
    },
    "/api/workflows/{id}/run": {
      "post": {
        "summary": "Run a workflow",
        "description": "Runs a workflow by its ID for the authenticated user.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow run successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Workflow not found."
          },
          "500": {
            "description": "Error running workflow."
          }
        }
      }
    },
...
  },
...
}